backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
sqlalchemy==2.0.23
pymysql==1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic-settings==2.3.4
python-dotenv==1.0.0
cryptography==41.0.7  # Required for MySQL auth